kmeans: use int instead of float for point attributes

+++ benchmarks/kmeans/bench.hpp
@@ -6,10 +6,10 @@ namespace kmeans {
 
 extern int numObjects;
 extern int numAttributes;
-extern float **attributes;
+extern int **attributes;
 extern int nclusters;
-extern float threshold;
-extern float **cluster_centres;
+extern int threshold;
+extern int **cluster_centres;
 
 #if !defined(USE_HB_MANUAL) && !defined(USE_HB_COMPILER)
 void run_bench(std::function<void()> const &bench_body,
@@ -20,11 +20,11 @@ void setup();
 void finishup();
 
 #if defined(USE_BASELINE)
-int cluster_serial(int numObjects, int numAttributes, float **attributes, int num_nclusters, float threshold, float ***cluster_centres);
+int cluster_serial(int numObjects, int numAttributes, int **attributes, int num_nclusters, int threshold, int ***cluster_centres);
 #elif defined(USE_OPENCILK)
-int cluster_opencilk(int numObjects, int numAttributes, float **attributes, int num_nclusters, float threshold, float ***cluster_centres);
+int cluster_opencilk(int numObjects, int numAttributes, int **attributes, int num_nclusters, int threshold, int ***cluster_centres);
 #elif defined(USE_OPENMP)
-int cluster_openmp(int numObjects, int numAttributes, float **attributes, int num_nclusters, float threshold, float ***cluster_centres);
+int cluster_openmp(int numObjects, int numAttributes, int **attributes, int num_nclusters, int threshold, int ***cluster_centres);
 #endif
 
 #if defined(TEST_CORRECTNESS)
